pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "kyc-automation"
version = "0.1.0"
description = "Multi-agent KYC exception handling automation using Microsoft Agent Framework"
requires-python = ">=3.11"
dependencies = [
    "agent-framework",
    "azure-identity",
    "azure-ai-formrecognizer>=3.3.0",
    "azure-cosmos>=4.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.setuptools.packages.find]
where = ["src"]
